help steps in a good place with dummy data
